backend/requirements.txt
fastapi
uvicorn
pandas
python-igraph
pydantic
python-multipart
